+++ singletons-th/CHANGES.md
@@ -4,6 +4,7 @@ Changelog for the `singletons-th` project
 next [????.??.??]
 -----------------
 * Require building with GHC 9.2.
+* Allow promoting and singling type applications in data constructor patterns.
 * Make the Template Haskell machinery generate `SingI1` and `SingI2` instances
   when possible.
 * Make `genDefunSymbols` and related functions less likely to trigger